+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [8.6.0] - 2025-09-25
+
+### Added
+
+- Add additional grouping discriminator property to events [#2544](https://github.com/bugsnag/bugsnag-js/pull/2544)
+- (react-native) Handle additional grouping discriminator [#2557](https://github.com/bugsnag/bugsnag-js/pull/2557)
+- (electron) Handle additional grouping discriminator [#2561](https://github.com/bugsnag/bugsnag-js/pull/2561)
+
+### Changed
+
+- Update bugsnag-cocoa to [v6.33.1](https//github.com/bugsnag/bugsnag-cocoa/releases/tag/v6.33.1) [#2552](https://github.com/bugsnag/bugsnag-js/pull/2552)
+- Update bugsnag-android to [v6.18.0](https//github.com/bugsnag/bugsnag-android/releases/tag/v6.18.0) [#2556](https://github.com/bugsnag/bugsnag-js/pull/2556)
+
+### Fixed
+
+- (plugin-window-onerror) Guard against errors when trying to call previous handler [#2551](https://github.com/bugsnag/bugsnag-js/pull/2551)
+
 ## [8.5.0] - 2025-09-01
 
 ### Added
@@ -38,7 +55,6 @@
 ## [8.3.0] - 2025-03-04
 
 This release adds support for React Native 0.77 and 0.78 to `@bugsnag/react-native`
-
 ### Added
 
 - (react-native-cli) Add support for `AppDelegate.swift` files to insert command [#2319](https://github.com/bugsnag/bugsnag-js/pull/2319)